--- src/main/java/server/Main.java
@@ -1,8 +1,0 @@
-package server;
-
-public class Main {
-    public static void main(String[] args) {
-        Server server = new Server();
-        server.launch();
-    }
-}
