+++ src/main/java/services/Service.java
@@ -1,0 +1,7 @@
+package services;
+
+public interface Service {
+
+    void launch();
+
+}
