+++ src/main/resources/META-INF/persistence.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<persistence xmlns="http://java.sun.com/xml/ns/persistence" version="2.0">
+
+    <persistence-unit name="Trust">
+        <class>model.User</class>
+        <properties>
+             <property name="javax.persistence.schema-generation.database.action" value="drop-and-create"/>
+             <property name="javax.persistence.jdbc.driver" value="com.mysql.jdbc.Driver" />
+            <property name="javax.persistence.jdbc.url" value="jdbc:mysql://mysql.server:3306/lisica.a_trust_test" />
+            <property name="javax.persistence.jdbc.user" value="lisica.a" />
+            <property name="javax.persistence.jdbc.password" value="Zxcvb123" />
+        </properties>
+    </persistence-unit>
+</persistence>
